+++ packages/perseus/src/widgets/categorizer/categorizer.tsx
@@ -14,7 +14,7 @@ import Renderer from "../../renderer";
 import mediaQueries from "../../styles/media-queries";
 import sharedStyles from "../../styles/shared";
 import Util from "../../util";
-import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/categorizer/prompt-utils";
+import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/categorizer/categorizer-ai-utils";
 
 import scoreCategorizer from "./score-categorizer";
 
@@ -24,7 +24,7 @@ import type {
     PerseusCategorizerRubric,
     PerseusCategorizerUserInput,
 } from "../../validation.types";
-import type {CategorizerPromptJSON} from "../../widget-ai-utils/categorizer/prompt-utils";
+import type {CategorizerPromptJSON} from "../../widget-ai-utils/categorizer/categorizer-ai-utils";
 
 type Props = WidgetProps<RenderProps, PerseusCategorizerRubric> & {
     values: ReadonlyArray<string>;
